backend/requirements.txt
Flask
psycopg2
python-decouple

''' Lista de Pacotes que precisa para rodar corretamente'''
